+++ kinto/core/openapi.py
@@ -28,7 +28,10 @@ class OpenAPI(CorniceSwagger):
         """Allow security extensions to expose authentication methods on the
         OpenAPI documentation. The definition field should correspond to a
         valid OpenAPI security definition. Refer the OpenAPI 2.0 specification
-        for more information. Below are some examples for BasicAuth and OAuth2::
+        for more information.
+        https://github.com/OAI/OpenAPI-Specification/blob/master/versions/2.0.md
+
+        Below are some examples for BasicAuth and OAuth2::
 
             {
                 "type": "basic",