+++ CHANGELOG.rst
@@ -6,6 +6,10 @@ This document describes changes between each past release.
 8.1.6 (unreleased)
 ------------------
 
+**New features**
+
+- Add Openid connect support (#939, #1425). See `demo <https://github.com/leplatrem/kinto-oidc-demo>`_
+
 **Bug fixes**
 
 - Fix missing principals from user info in root URL when default bucket plugin is enabled (fixes #1495)
@@ -22,7 +26,7 @@ This document describes changes between each past release.
   that we can still fail to INSERT after the check succeeded, but it
   can reduce write load in configurations where there are a lot of
   create()s (i.e. when using the default_bucket plugin).
-
+ 
 
 8.1.4 (2018-01-31)
 ------------------
@@ -31,6 +35,7 @@ This document describes changes between each past release.
 
 - Allow inherited resources to set a custom model instance before instantiating (fixes #1472)
 - Fix collection timestamp retrieval when the stack is configured as readonly (fixes #1474)
+
 
 8.1.3 (2018-01-26)
 ------------------